+++ governance/engine.go
@@ -55,6 +55,9 @@ type Markets interface {
 	MarketExists(market string) bool
 	GetMarket(market string) (types.Market, bool)
 	GetEquityLikeShareForMarketAndParty(market, party string) (num.Decimal, bool)
+	RestoreMarket(ctx context.Context, marketConfig *types.Market) error
+	RestoreMarketWithLiquidityProvision(ctx context.Context, marketConfig *types.Market, lp *types.LiquidityProvisionSubmission, lpid, deterministicID string) error
+	StartOpeningAuction(ctx context.Context, marketID string) error
 }
 
 // StakingAccounts ...
@@ -491,10 +494,9 @@ func (e *Engine) intoToSubmit(ctx context.Context, p *types.Proposal) (*ToSubmit
 		// FIXME(): normally we should use the closetime
 		// but this would not play well with the MarketAuctionState stuff
 		// for now we start the auction as of now.
-		closeTime := e.timeService.GetTimeNow()
+		closeTime := e.timeService.GetTimeNow().Truncate(time.Second)
 		enactTime := time.Unix(p.Terms.EnactmentTimestamp, 0)
 		newMarket := p.Terms.GetNewMarket()
-
 		auctionDuration := enactTime.Sub(closeTime)
 		if perr, err := validateNewMarketChange(newMarket, e.assets, true, e.netp, auctionDuration); err != nil {
 			e.rejectProposal(ctx, p, perr, err)